refactor : 필터된 레스토랑 카드를 렌더링 하는 함수 구분
