+++ src/components/FilterSelector/NameOrDistanceSelector/index.js
@@ -1,7 +1,7 @@
 import changeModalContents from "../../../changeModalContents";
 import storage from "../../../domain/storage";
+import renderFilteredRestaurants from "../../../renderFilteredRestaurant";
 import createRestaurantCards from "../../../service/createRestaurantCards";
-import renderRestaurants from "../../../ui/renderRestaurant";
 import createKeyValuePair from "../../../utils/createKeyValuePair";
 import { $ } from "../../../utils/dom";
 import Select from "../../common/Select";
@@ -11,19 +11,7 @@ const NameOrDistanceSelector = (restaurantList) => {
     change: (e) => {
       storage.saveNameOrDistance(e.target.value);
       restaurantList.setNameOrDistance(e.target.value);
-      renderRestaurants(
-        createRestaurantCards(restaurantList.filter(), {
-          clickCard: (restaurant) => {
-            $("#restaurant-detail-modal-backdrop").classList.add("open");
-            changeModalContents(restaurant, restaurantList);
-          },
-          clickFavorite: () => {
-            storage.saveRestaurantList(
-              restaurantList.list.map((restaurant) => restaurant.value)
-            );
-          },
-        })
-      );
+      renderFilteredRestaurants(restaurantList);
     },
   };
