+++ src/main.js
@@ -3,14 +3,12 @@ import header from "./components/Header";
 import Modal from "./components/common/Modal";
 import Title from "./components/common/Title";
 import RegisterForm from "./components/RegisterForm";
-import renderRestaurants from "./ui/renderRestaurant";
 import { registerModalClose } from "./components/common/Modal/registerModalClose";
 import Restaurant from "./domain/Restaurant";
 import RestaurantList from "./domain/RestaurantList";
 import CategorySelector from "./components/FilterSelector/CategorySelector";
 import NameOrDistanceSelector from "./components/FilterSelector/NameOrDistanceSelector";
 import Tab from "./components/Tab";
-import renderAllpage from "./ui/renderAllpage";
 import storage from "./domain/storage";
 import { restaurants } from "./restaurantListData";
 import renderFavoritePage from "./ui/renderFavoritePage";
